:sparkles: 新增分片任务

+++ lingting-core/src/main/java/live/lingting/framework/util/FileUtils.java
@@ -187,7 +187,9 @@ public class FileUtils {
 
 	public static boolean delete(File file) {
 		try {
-			Files.delete(file.toPath());
+			if (file != null) {
+				Files.delete(file.toPath());
+			}
 			return true;
 		}
 		catch (IOException e) {
